+++ fs/inode/inode.h
@@ -390,7 +390,7 @@ int files_allocate(FAR struct inode *inode, int oflags, off_t pos,
                    FAR void *priv, int minfd);
 
 /****************************************************************************
- * Name: files_dupfd2
+ * Name: files_dup2
  *
  * Description:
  *   Clone a file descriptor to a specific descriptor number.
@@ -401,7 +401,7 @@ int files_allocate(FAR struct inode *inode, int oflags, off_t pos,
  *
  ****************************************************************************/
 
-int files_dupfd2(int fd1, int fd2);
+int files_dup2(int fd1, int fd2);
 
 /****************************************************************************
  * Name: files_close
